+++ terraform/destroy.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-terraform destroy -auto-approve
-echo -e "${GREEN}The cluster was successfully destroyed by Terraform.${NORMAL}"