+++ terraform/kubernetes/gke.tf
@@ -3,8 +3,6 @@ resource "google_container_cluster" "primary" {
   name     = "${var.project_id}-gke"
   location = var.zone
 
-  enable_legacy_abac = true
-
   remove_default_node_pool = true
   initial_node_count       = var.gke_num_nodes